requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

# Core ML / image
numpy==1.23.5
opencv-python==4.7.0.72
scikit-image==0.21.0
pillow==9.5.0
onnx==1.14.0
onnxruntime-gpu==1.16.2
protobuf==4.23.2
torch==2.0.1+cu118
torchvision==0.15.2
torchaudio==2.0.2
tqdm
ftfy
regex

# Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
msgpack>=1.0.7
websockets>=12.0
python-dotenv>=1.0.0

# Client (face detection — CPU-only, no GPU needed on client machine)
mediapipe>=0.10.9
